user-guide: Add firewall section
Add a section about firewall to the user guide. This guide also includes
about the firewall templates. This is useful for helping the users
understand the firewall, and how to configure it.

Also, add the automatic generation of section labels, as the
SULKA_NFTABLES_CONF variable description refers to the firewall section
in the same page.